Fix model and associations

+++ Codigo/dindin-api/dindin-app/app/database/migrations/20220420143208-create-category.js
@@ -23,7 +23,7 @@ module.exports = {
                     // ! },
                 },
                 user_id: {
-                    type: DataTypes.INTEGER(11).UNSIGNED,
+                    type: DataTypes.INTEGER.UNSIGNED,
                     allowNull: false,
                     references: {
                         model: {
